# Backend/docs/swagger.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Library-API",
        "version": "1.0.0",
        "description": "### This is a Swagger UI for our Library API  \n\n### Error handling\nErrors are presented using the HTTP status codes 400-599. The response include a more detailed explanation about the error in the body.\n\n```\n{\n  \"error\": \"Cleartext explanation of the error that occurred\"\n}\n```\n\n### Generic API errors\nThese error responses can be returned by any of the described endpoints.\n\n| HTTP status  |  Description |\n|---|---|\n| 400| Bad input parameters or invalid request. More information can be found in the response body.\n| 401| The consumer does not have access to the requested resource or the provided token is not valid.\n| 404| The requested resource can not be found or the user does not have access to the resource.\n| 5XX| A server error has occurred while processing the request."
    },
    "paths": {
...
        "/books/author/:author": {
            "get": {
                "tags": [
                    "books"
                ],
                "parameters": [
                    {
                        "examples": {
                            "key": {
                                "value": "30cf8c1d-3fb9-4149-a2eb-20e3e88b89d6"
                            }
                        },
                        "name": "key",
                        "description": "API key is needed to access the API",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    },
                    {
                        "examples": {
                            "author": {
                                "value": "John Ronald Reuel Tolkien"
                            }
                        },
                        "name": "author",
                        "description": "Name of an author whos books to get",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                },
                                "examples": {
                                    "Books": {
                                        "value": [
                                            {
                                                "_id": "63186aac214030a84b0223e8",
                                                "title": "The Lord of the Rings",
                                                "author": "John Ronald Reuel Tolkien",
                                                "stock": 6,
                                                "imgSource": "http://localhost:3000/bookCovers/coverImg-1663582105444lord-of-the-rings.jpg"
                                            },
                                            {
                                                "_id": "6320516964da52b099469bb0",
                                                "title": "The Hobbit",
                                                "author": "John Ronald Reuel Tolkien",
                                                "stock": 4,
                                                "imgSource": "http://localhost:3000/bookCovers/coverImg-1663582105444the-hobbit.jpg"
                                            }
                                        ]
                                    }
                                }
                            }
                        },
                        "description": "Request successful"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Missing API key error": {
                                        "value": {
                                            "error": "Missing API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is missing"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Invalid API key error": {
                                        "value": {
                                            "error": "Invalid API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is invalid"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Author not found error": {
                                        "value": {
                                            "error": "Looks like we couldn't find what you were looking for."
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Author not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Something went wrong error": {
                                        "value": {
                                            "error": "Looks like something went wrong :("
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Something went wrong during the query"
                    }
                },
                "summary": "Get a list of books by an author"
            }
        },
        "/books/genre/:genre": {
            "get": {
                "tags": [
                    "books"
                ],
                "parameters": [
                    {
                        "examples": {
                            "key": {
                                "value": "30cf8c1d-3fb9-4149-a2eb-20e3e88b89d6"
                            }
                        },
                        "name": "key",
                        "description": "API key is needed to access the API",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    },
                    {
                        "examples": {
                            "genre": {
                                "value": "fantasy"
                            }
                        },
                        "name": "genre",
                        "description": "genre to find a book",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                },
                                "examples": {
                                    "Books": {
                                        "value": [
                                            {
                                                "id": "43f7d8e0-2e93-11ed-9a62-ad6ce2fced68",
                                                "title": "The Lord of the Rings",
                                                "author": "John Ronald Reuel Tolkien",
                                                "stock": 6,
                                                "imgSource": "http://localhost:3000/bookCovers/coverImg-1663582105444lord-of-the-rings.jpg"
                                            },
                                            {
                                                "_id": "6320516964da52b099469bb0",
                                                "title": "The Hobbit",
                                                "author": "John Ronald Reuel Tolkien",
                                                "stock": 4,
                                                "imgSource": "http://localhost:3000/bookCovers/coverImg-1663582105444the-hobbit.jpg"
                                            }
                                        ]
                                    }
                                }
                            }
                        },
                        "description": "Request successful"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Missing API key error": {
                                        "value": {
                                            "error": "Missing API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is missing"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Invalid API key error": {
                                        "value": {
                                            "error": "Invalid API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is invalid"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Genre not found error": {
                                        "value": {
                                            "error": "Looks like we couldn't find what you were looking for."
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Category not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Something went wrong error": {
                                        "value": {
                                            "error": "Looks like something went wrong :("
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Something went wrong during the query"
                    }
                },
                "summary": "Get a list of books by genre"
            }
        },
...
        "/reviews/book/:id": {
            "get": {
                "tags": [
                    "reviews"
                ],
                "parameters": [
                    {
                        "examples": {
                            "key": {
                                "value": "30cf8c1d-3fb9-4149-a2eb-20e3e88b89d6"
                            }
                        },
                        "name": "key",
                        "description": "API key is required to access the API",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    },
                    {
                        "examples": {
                            "id": {
                                "value": "63283f9913c575a0084104ee"
                            }
                        },
                        "name": "id",
                        "description": "A id of a book",
                        "schema": {
                            "type": "string"
                        },
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                },
                                "examples": {
                                    "Reviews": {
                                        "value": [
                                            {
                                                "_id": "6328614b761359e66be3f007",
                                                "title": "test",
                                                "body": "The missile knows where it is at all times. It knows this because it knows where it isn't, by subtracting where it is, from where it isn't, or where it isn't, from where it is, whichever is greater, it obtains a difference, or deviation. The guidance sub-system uses deviations to generate corrective commands to drive the missile from a position where it is, to a position where it isn't, and arriving at a position where it wasn't, it now is. Consequently, the position where it is, is now the position that it wasn't, and it follows that the position where it was, is now the position that it isn't. In the event of the position that it is in is not the position that it wasn't, the system has required a variation. The variation being the difference between where the missile is, and where it wasn't. If variation is considered to be a significant factor, it too, may be corrected by the GEA. However, the missile must also know where it was. The missile guidance computance scenario works as follows: Because a variation has modified some of the information the missile has obtained, it is not sure just where it is, however it is sure where it isn't, within reason, and it knows where it was. It now subracts where it should be, from where it wasn't, or vice versa. By differentiating this from the algebraic sum og where it shouldn't be, and where it was. It is able to obtain a deviation, and a variation, which is called \"air\"",
                                                "rating": 1,
                                                "book": "63283f9913c575a0084104ee",
                                                "author": "63285d6a761359e66be3efe7",
                                                "createdAt": 1663589144454
                                            },
                                            {
                                                "_id": "632869769d0f2098cf6bd58b",
                                                "title": "this book is the second coming of christ",
                                                "body": "There was no ring on his finger. That was a good sign although far from proof that he was available. Still, it was much better than if he had been wearing a wedding ring on his hand. She glanced at his hand a bit more intently to see if there were any tan lines where a ring may have been, and he's simply taken it off. She couldn't detect any which was also a good sign and a relief. The next step would be to get access to his wallet to see if there were any family photos in it.",
                                                "rating": 10,
                                                "book": "63283f9913c575a0084104ee",
                                                "author": "63285d6a761359e66be3efe7",
                                                "createdAt": 1663591492087
                                            }
                                        ]
                                    }
                                }
                            }
                        },
                        "description": "Request successful"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Missing API key error\t": {
                                        "value": {
                                            "error": "Missing API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is not provided"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Invalid API key error": {
                                        "value": {
                                            "error": "Invalid API key"
                                        }
                                    }
                                }
                            }
                        },
                        "description": "API key is invalid"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unable to find reviews": {
                                        "value": {
                                            "error": "Looks like we couldn't find what you were looking for."
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Review was not found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Something went wrong error": {
                                        "value": {
                                            "error": "Looks like something went wrong :("
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Something went wrong during the query"
                    }
                },
                "summary": "Get a specific books reviews"
            }
        },
...
    }
}